+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # 0.9.6+ - [Unreleased]
 ## Added
+- `--quiet` option
 
 ## Fixed